data/cloud_mldeployment/sentiment_analysis/requirements.txt
azure_storage==0.37.0
joblib==1.3.1
pandas==2.0.3
python-dotenv==1.0.0
scikit_learn==1.3.0
